Fix Checkstyle and refactor to booking
